fix(EE): LW must sign-extend into the 64-bit GPR (was zero-extending) — root cause of the real boot's infinite loop
The single highest-impact bug found this session. ExecuteLw did
`SetGpr(rt, new Gpr128 { Lo = _memory.Read32(addr) })` — assigning a
uint to a ulong field is a zero-extending widening conversion in C#.
Real MIPS III/R5900 LW is architecturally defined to sign-extend the
loaded 32-bit value into the 64-bit register — that's precisely what
distinguishes plain LW from LWU (load word unsigned), which this
codebase already correctly zero-extends. Loading 0xFFFFFFFF via LW was
producing +4,294,967,295 in the 64-bit register instead of the correct
-1.

Found by hand-disassembling the real boot's stall (using the new
--trace-window CLI diagnostic to get the loop's exact, minimal
instruction footprint) down to a single instruction: `slt v0,s0,a1`
comparing a small positive counter against a1, which had just been
loaded via LW from a stack slot holding 0xFFFFFFFF. Real signed MIPS
semantics: `positive < -1` is false, so the loop's `bne v0,zero` should
never re-take the branch — the loop should exit on its first check.
Instead, because of the zero-extension bug, the comparison was
`positive < +4294967295`, always true, so the loop never terminated.
This is exactly the "for (i=0; i<(unsigned)-1; i++) { ...; if (done)
break; }" idiom — a completely ordinary, correct pattern in the game's
own code that our LW bug turned into a true infinite loop.

This isn't specific to this one loop or this one title: LW is one of
the single most common instructions in any compiled MIPS program, and
the sign-vs-zero-extension distinction matters every time a loaded
32-bit value has its high bit set and gets used in any 64-bit-width
context — comparisons, pointer arithmetic, sign-sensitive branches.
Likely one of the most broadly consequential correctness bugs in the
whole emulator.

Verified: full smoke suite green (no regressions from a change this
foundational). Real-boot effect is dramatic — the observed infinite
loop (previously running for 6+ billion cycles / 20+ real seconds with
zero forward progress) is completely gone; execution now reaches
entirely new code and a different, later point in the boot sequence
before its next stall (a separate issue, not yet root-caused).

+++ src/DetPS2.Core/EmotionEngine.cs
@@ -874,7 +874,14 @@ public sealed class EmotionEngine : ISchedulable
         uint rs = (opcode >> 21) & 0x1F; uint rt = (opcode >> 16) & 0x1F;
         short off = (short)(opcode & 0xFFFF);
         ulong addr = GetGpr(rs).Lo + (ulong)off;
-        if (rt != 0) SetGpr(rt, new Gpr128 { Lo = _memory.Read32(addr) });
+        // LW sign-extends into the 64-bit GPR (that's exactly what distinguishes it from
+        // LWU, which zero-extends) — a plain uint->ulong assignment here is a zero-extend,
+        // silently turning any loaded value with the high bit set (e.g. -1, or any negative
+        // 32-bit int/loop-bound-sentinel) into a huge positive 64-bit number instead. Found
+        // by hand-tracing a real infinite loop: `slt v0,s0,a1` with a1 loaded as 0xFFFFFFFF
+        // via LW should compare as "positive < -1" (false, loop should end) but was instead
+        // comparing "positive < +4294967295" (true, forever) because of this bug.
+        if (rt != 0) SetGpr(rt, new Gpr128 { Lo = (ulong)(long)(int)_memory.Read32(addr) });
     }
 
     private void ExecuteSb(uint opcode)
